requirements.txt
absl-py==2.3.1
actionlib==1.14.3
angles==1.9.14
annotated-types==0.7.0
base_local_planner==1.17.3
camera-calibration-parsers==1.12.1
catkin==0.8.12
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cv-bridge==1.16.2
cycler==0.12.1
diagnostic-updater==1.12.1
dynamic-reconfigure==1.7.6
Farama-Notifications==0.0.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.76.0
gym==0.26.2
gym-notices==0.1.0
gymnasium==1.2.2
HeapDict==1.0.1
idna==3.11
interactive-markers==1.12.2
Jinja2==3.1.6
joint-state-publisher==1.15.2
kiwisolver==1.4.9
laser-geometry==1.6.8
loguru==0.7.3
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
message-filters==1.17.4
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
platformdirs==4.5.1
protobuf==6.33.1
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-qt-binding==0.4.6
pytz==2025.2
PyYAML==6.0.3
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
requests==2.32.5
resource-retriever==1.12.10
rosbag==1.17.4
rosclean==1.15.10
rosgraph==1.17.4
roslaunch==1.17.4
roslib==1.15.10
roslz4==1.17.4
rosmaster==1.17.4
rosmsg==1.17.4
rosnode==1.17.4
rosparam==1.17.4
rospy==1.17.4
rosservice==1.17.4
rostest==1.17.4
rostopic==1.17.4
rosunit==1.15.10
roswtf==1.17.4
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-tf-tree==0.6.5
rviz==1.14.26
scipy==1.16.3
seaborn==0.13.2
sensor-msgs==1.13.2
sentry-sdk==2.47.0
setuptools==80.9.0
Shimmy==2.0.0
six==1.17.0
smclib==1.8.7
smmap==5.0.2
sortedcontainers==2.4.0
stable_baselines3==2.7.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tf==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
tf2-sensor-msgs==0.7.10
topic-tools==1.17.4
torch==2.9.1
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.1
wandb==0.23.1
Werkzeug==3.1.4
wheel==0.45.1
xacro==1.14.20
